first version to claim rewards
