+++ phat/contracts/dapp_staking/.env_shibuya
@@ -9,9 +9,3 @@ CONTRACT_ID=5b665bf5f5fac00d3201faa326255df76343720e5385b02492554bd9a8889b8b
 # ATTEST_KEY=e5be9a5092b81bca64be81d212e7f2f9eba183bb7a90954f7b76361f6edb5c0a
 # Enable Meta-Tx by uncommenting the next line. It uses //bob
 # SENDER_KEY=398f0c28f98885e046333d4a41c19cee4c37368a9832c6502f6cfd182e2aef89
-
-GRAPH_API=https://query.substrate.fi/dapps-staking-subquery-astar
-DAPP_ID=zsv1gvepvmwfdshmwgczs4zyvmmwesbjwqjn4wdpuefrrpy
-
-
-0xe068747470733a2f2f71756572792e7375627374726174652e66692f64617070732d7374616b696e672d73756271756572792d617374617276bc7a73763167766570766d77666473686d7767637a73347a79766d6d776573626a77716a6e3477647075656672727079
