+++ phat/contracts/dapp_staking/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "js_offchain_rollup"
+name = "dapp_staking"
 version = "0.1.0"
 authors = ["GuiGou"]
 edition = "2021"
@@ -19,8 +19,6 @@ fixed = { version = "1", default-features = false, features = ["serde"] }
 pink-extension = { version = "0.4.2", default-features = false }
 pink-json = { git = "https://github.com/Phala-Network/pink-json.git", branch = "pink", default-features = false, features = ["de-number-as-str"] }
 pink-web3 = { version = "0.20.2", default-features = false, features = ["pink", "signing"] }
-
-phat_offchain_rollup = { git = "https://github.com/GuiGou12358/phat-offchain-rollup" , default-features = false, features = ["ink", "logging"] }
 
 subrpc = { package = "pink-subrpc", version = "0.4.2", default-features = false }
 hex = { version = "0.4", default-features = false }
@@ -46,7 +44,6 @@ std = [
     "ink/std",
     "scale/std",
     "scale-info/std",
-    "phat_offchain_rollup/std",
     "pink-extension/std",
     "subrpc/std",
     "pink-json/std",
@@ -54,4 +51,3 @@ std = [
     "phat_js/std",
 ]
 ink-as-dependency = []
-logging = ["phat_offchain_rollup/logging"]
